Ref: refactored whole study page #75
