+++ src/components/study/dashboard/admin/ApplyResult.jsx
