+++ src/components/ui/nav/StudyDashboardNav.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
-export default function Navigation({ isAdminPage = false, currentMenu, onMenuClick }) {
+export default function StudyDashboardNav({ isAdminPage = false, currentMenu, onMenuClick }) {
     const myMenuItems = [
         { id: 'applyResult', label: '스터디 지원 결과' },
         { id: 'attendance', label: '스터디 출석기록' },
